Add landing page and move profile
